fix(examples): Move examples out of fake dependency, mainly on Geant3
Made order in the dependcies of many of the examples.
Most were placed in the `if(Geant3_FOUND)` which is not needed.
Moved `propagator` out of `if(TARGET FairRoot::EventDisplay`
following previous commit.

That increased the number of tests for the no-Geant3 case.

(cherry picked from commit c185b1d0d9cc5c821f22ca68ae471241ad13538a)